+++ pkg/mgmt/builder.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package mgmt
 
 import (
 	"github.com/Sirupsen/logrus"
