+++ pkg/mgmt/start.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package mgmt
 
 import (
 	"sync"
